update routes and core
