+++ services/backend/src/core/admin/users.js
@@ -1,5 +1,5 @@
 const _ = require("lodash");
-const prisma = require("../../../database");
+const prisma = require("../../database");
 
 const getUsers = async (request, response) => {
   const { language } = request.query;
