+++ services/backend/src/core/options/careerMobilities.js
@@ -1,5 +1,5 @@
 const _ = require("lodash");
-const prisma = require("../../../database");
+const prisma = require("../../database");
 
 async function getCareerMobilities(request, response) {
   const { language } = request.query;
