+++ services/backend/src/docs/swaggerOptions.js
@@ -45,8 +45,6 @@ module.exports = swaggerUi.setup(specs, {
     defaultModelsExpandDepth: -1,
     filter: true,
     withCredentials: true,
-    tagsSorter: (a, b) => {
-      return a.localeCompare(b);
-    },
+    tagsSorter: (a, b) => a.localeCompare(b),
   },
 });
