+++ services/backend/src/router/admin/admin.js
@@ -1,9 +1,9 @@
 const { Router } = require("express");
 const { keycloak } = require("../../auth/keycloak");
-const admin = require("../../core/admin/admin");
+const admin = require("../../core/admin/users");
 const { updateUserStatusValidator } = require("./validator");
 const { langValidator } = require("../util/commonValidators");
-const { validationMiddlware } = require("../../utils/middlewares");
+const { validationMiddleware } = require("../../utils/middleware");
 
 const adminRouter = Router();
 
@@ -11,7 +11,7 @@ adminRouter.get(
   "/users",
   keycloak.protect("view-admin-console"),
   langValidator,
-  validationMiddlware,
+  validationMiddleware,
   admin.getUsers
 );
 
@@ -19,7 +19,7 @@ adminRouter.put(
   "/userStatuses",
   keycloak.protect("manage-users"),
   updateUserStatusValidator,
-  validationMiddlware,
+  validationMiddleware,
   admin.updateUserStatuses
 );
