+++ services/backend/src/router/admin/validator.js
@@ -3,12 +3,12 @@ const { isUUID, isIn } = require("validator");
 
 const updateUserStatusValidator = [
   body()
-    .custom((object) => {
-      return Object.keys(object).every(
+    .custom((object) =>
+      Object.keys(object).every(
         (key) =>
           isUUID(key) && isIn(object[key], ["ACTIVE", "INACTIVE", "HIDDEN"])
-      );
-    })
+      )
+    )
     .withMessage(
       "must be a JSON object with user UUIDs as keys and it's value must be 'ACTIVE' or 'INACTIVE' or 'HIDDEN'"
     ),
