+++ services/backend/src/router/connections/connections.js
@@ -4,13 +4,13 @@ const { keycloak } = require("../../auth/keycloak");
 const connections = require("../../core/connections/connections");
 
 const { UUIDValidator } = require("../util/commonValidators");
-const { validationMiddlware } = require("../../utils/middlewares");
+const { validationMiddleware } = require("../../utils/middleware");
 
 const connectionsRouter = Router();
 
 connectionsRouter
   .route("/:id")
-  .all(keycloak.protect(), UUIDValidator, validationMiddlware)
+  .all(keycloak.protect(), UUIDValidator, validationMiddleware)
   .get(connections.getConnectionById)
   .post(connections.addConnection)
   .delete(connections.removeConnection);
