+++ services/backend/src/router/geds/geds.js
@@ -2,7 +2,7 @@ const { Router } = require("express");
 const { keycloak } = require("../../auth/keycloak");
 const profileGen = require("../../core/geds/geds");
 const { profileGenValidator } = require("./validator");
-const { validationMiddlware } = require("../../utils/middlewares");
+const { validationMiddleware } = require("../../utils/middleware");
 
 const profileGenRouter = Router();
 
@@ -10,7 +10,7 @@ profileGenRouter.get(
   "/",
   keycloak.protect(),
   [profileGenValidator],
-  validationMiddlware,
+  validationMiddleware,
   profileGen.getGedsSetup
 );
