+++ services/backend/src/router/search/search.js
@@ -3,7 +3,7 @@ const { fuzzyValidator } = require("./validator");
 const { keycloak } = require("../../auth/keycloak");
 const { filterSearch, fuzzySearch } = require("../../core/search/search");
 const { langValidator } = require("../util/commonValidators");
-const { validationMiddlware } = require("../../utils/middlewares");
+const { validationMiddleware } = require("../../utils/middleware");
 
 const searchRouter = Router();
 
@@ -11,14 +11,14 @@ searchRouter.get(
   "/filters",
   keycloak.protect(),
   langValidator,
-  validationMiddlware,
+  validationMiddleware,
   filterSearch
 );
 searchRouter.get(
   "/fuzzy",
   keycloak.protect(),
   [langValidator, fuzzyValidator],
-  validationMiddlware,
+  validationMiddleware,
   fuzzySearch
 );
