+++ services/backend/src/server.js
@@ -11,27 +11,27 @@ const { keycloak, sessionInstance } = require("./auth/keycloak");
 const router = require("./router/router");
 const swaggerOptions = require("./docs/swaggerOptions");
 const config = require("./config");
-const { errorHandler } = require("./utils/middlewares");
+const { errorHandler } = require("./utils/middleware");
 
 const app = express();
 
-app.set("trust proxy", true);
-app.use((req, res, next) => {
-  res.header("Access-Control-Allow-Origin", "*");
-  res.header(
-    "Access-Control-Allow-Headers",
-    "Origin, Authorization, X-Requested-With, Content-Type, Accept"
-  );
-  res.header("Access-Control-Allow-Methods", "PUT, POST, GET, DELETE, OPTIONS");
-  next();
-});
+// app.set("trust proxy", true);
+// app.use((req, res, next) => {
+//   res.header("Access-Control-Allow-Origin", "*");
+//   res.header(
+//     "Access-Control-Allow-Headers",
+//     "Origin, Authorization, X-Requested-With, Content-Type, Accept"
+//   );
+//   res.header("Access-Control-Allow-Methods", "PUT, POST, GET, DELETE, OPTIONS");
+//   next();
+// });
 
 if (config.ENV !== "test") {
   app.use(morgan(config.MORGAN_CONFIG));
 }
 
 if (config.ENV === "development") {
-  app.get("/oauth2-redirect.html", function (req, res) {
+  app.get("/oauth2-redirect.html", (req, res) => {
     res.sendFile(`${__dirname}/docs/oauth2-redirect.html`);
   });
   app.use("/api-docs", swaggerUi.serve, swaggerOptions);
@@ -42,7 +42,6 @@ app.use(
   helmet(),
   sessionInstance,
   timeout("5s"),
-  keycloak.middleware(),
   bodyParser.urlencoded({ extended: true }),
   bodyParser.json()
 );
